EZP-23940: fixed: test case classes namespaces not updated after move
